Feature/lab5 v2/victor (#16)
* basic authentication and authorization

* add role to accessToken

* create all tables for database and configure them + make logic onsisntent and refactor

---------

+++ PedagogyPrime/PedagogyPrime.Infrastructure/Commands/Courses/Delete/DeleteCourseCommandHandler.cs
@@ -1,25 +1,35 @@
-﻿using MediatR;
-using PedagogyPrime.Core.Common;
+﻿using PedagogyPrime.Core.Common;
 using PedagogyPrime.Core.IRepositories;
+using PedagogyPrime.Infrastructure.IAuthorization;
 
 namespace PedagogyPrime.Infrastructure.Commands.Courses.Delete
 {
-	public class DeleteCourseCommandHandler : IRequestHandler<DeleteCourseCommand, BaseResponse<bool>>
+	using Common;
+
+	public class DeleteCourseCommandHandler : BaseRequestHandler<DeleteCourseCommand, BaseResponse<bool>>
 	{
 		private readonly ICourseRepository courseRepository;
 
-		public DeleteCourseCommandHandler(ICourseRepository courseRepository)
+		public DeleteCourseCommandHandler(
+			ICourseRepository courseRepository,
+			IUserAuthorization userAuthorization
+		) : base(userAuthorization)
 		{
 			this.courseRepository = courseRepository;
 		}
 
-		public async Task<BaseResponse<bool>> Handle(
+		public override async Task<BaseResponse<bool>> Handle(
 			DeleteCourseCommand request,
 			CancellationToken cancellationToken
 		)
 		{
 			try
 			{
+				if(!(await IsAuthorized(request.UserId)))
+				{
+					return BaseResponse<bool>.Forbbiden();
+				}
+
 				var result = await courseRepository.Delete(request.Id);
 
 				if(result == 0)
